Reorganization of the tools
