+++ src/main/resources/application.conf
@@ -26,9 +26,10 @@ bss {
   root-path = "."
   root-path = ${?BSS_ROOT_PATH}
 
-  watcher {
-    db-path = "./dbs/watcher"
-    db-path = ${?BSS_WATCHER_DB_PATH}
-  }
+  db-base-path = "./db"
+  db-base-path = ${?BSS_DB_BASE_PATH}
 
+  watcher {
+    db-path = ${bss.db-base-path}/watcher
+  }
 }
